Move ChecksSelection to pages
