+++ assets/js/pages/ChecksSelection/ChecksSelectionGroup.test.jsx
