+++ assets/js/pages/ChecksSelection/ChecksSelectionItem.jsx
