+++ assets/js/pages/ChecksSelection/index.js
